define event name
delete define.go
